# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: "0.2.0"
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /push/subscribe:
    post:
      operationId: subscribeToPush
      tags: [push]
      summary: Register a browser push subscription for the current user
      description: >
        Upserted by endpoint — re-subscribing the same browser updates its
        keys rather than creating a duplicate. A user can have several
        (one per browser/device).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PushSubscriptionInput"
      responses:
        "200":
          description: Subscribed
    delete:
      operationId: unsubscribeFromPush
      tags: [push]
      summary: Remove a browser push subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnsubscribePushInput"
      responses:
        "200":
          description: Unsubscribed (or was already not subscribed)
...
components:
  schemas:
...
    PushSubscriptionInput:
      type: object
      required: [endpoint, keys]
      properties:
        endpoint:
          type: string
          minLength: 1
        keys:
          type: object
          required: [p256dh, auth]
          properties:
            p256dh:
              type: string
              minLength: 1
            auth:
              type: string
              minLength: 1
    UnsubscribePushInput:
      type: object
      required: [endpoint]
      properties:
        endpoint:
          type: string
          minLength: 1
...