# lib/api-zod/src/generated/types/pushSubscriptionInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { PushSubscriptionInputKeys } from './pushSubscriptionInputKeys';

export interface PushSubscriptionInput {
  /** @minLength 1 */
  endpoint: string;
  keys: PushSubscriptionInputKeys;
}